move overlay to top
